undersync: move client module into separate package
